+++ CHANGELOG.md
@@ -10,6 +10,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ### Fixed
 
 - "@daily-co/daily-js" should not be included in the `rtvi-client-js` package.json. This dependency is only necessary for `rtvi-client-js-daily`.
+- Jest unit tests added for config manipulation within `rtvi-client-js` (`yarn run test`)
 
 ## [0.1.6] - 2024-08-26